requirements.txt
# FastAPI e servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Autenticação e segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Variáveis de ambiente
python-dotenv==1.0.0

# Validação e serialização
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Background tasks
celery==5.3.4
redis==5.0.1

# Logging
python-json-logger==2.0.7

# Testes (opcional para desenvolvimento)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
requests==2.31.0

